ml-models/sputum_detector/requirements.txt
shapely
tensorflow==2.11.0
opencv-python
numpy
mlflow==2.2.1
cloudpickle
requests
packaging
typer
fastapi
ml_collections
python-dotenv